Planar Straight Line Graph

Create a 'pslg' which is a mesh of segments used in the RTriangle package.

as_pslg(x)

# S3 method for default
as_pslg(x)

Arguments

x

data model (understood by SC0())

Value

object of class RTriangle::pslg from the RTriangle package

Details

The pslg is constructed from from unique vertices x, y and their segments. In the context of triangulating the functions DEL0() and DEL() do more to ensure that input polygons have their holes culled out (or classified by invisible triangles.)

Examples

data("minimal_mesh", package = "silicate")
as_pslg(minimal_mesh)
#> $P
#>         x_   y_
#>  [1,] 0.00 0.00
#>  [2,] 0.00 1.00
#>  [3,] 0.20 0.20
#>  [4,] 0.20 0.40
#>  [5,] 0.30 0.60
#>  [6,] 0.50 0.20
#>  [7,] 0.50 0.40
#>  [8,] 0.50 0.70
#>  [9,] 0.69 0.00
#> [10,] 0.75 1.00
#> [11,] 0.80 0.60
#> [12,] 1.00 0.80
#> [13,] 1.10 0.63
#> [14,] 1.23 0.30
#> 
#> $PA
#>      
#>  [1,]
#>  [2,]
#>  [3,]
#>  [4,]
#>  [5,]
#>  [6,]
#>  [7,]
#>  [8,]
#>  [9,]
#> [10,]
#> [11,]
#> [12,]
#> [13,]
#> [14,]
#> 
#> $PB
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $S
#>       .vx0 .vx1
#>  [1,]    1    2
#>  [2,]    2   10
#>  [3,]   10   12
#>  [4,]   12    8
#>  [5,]    8   11
#>  [6,]   11    9
#>  [7,]    9    1
#>  [8,]    3    6
#>  [9,]    6    7
#> [10,]    7    5
#> [11,]    5    4
#> [12,]    4    3
#> [13,]    9   11
#> [14,]   11   13
#> [15,]   13   14
#> [16,]   14    9
#> 
#> $SB
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $H
#>      [,1] [,2]
#> 
#> attr(,"class")
#> [1] "pslg"