Geocentric (XYZ) coordinates

Convert longitude/latitude coordinates to geocentric coordinates.

globe(x, ...)

# S3 method for mesh3d
globe(x, gproj = NULL, ...)

# S3 method for default
globe(x, gproj = "+proj=geocent +datum=WGS84", ...)

Arguments

x

a silicate model or mesh3d

...

arguments to methods (none used)

gproj

Geocentric PROJ.4 string, defaults to WGS84

Value

object with vertices table modified

Details

With silicate data checks are made for the projection in use, but not for mesh3d. In that case data are assumed to be 'longitude,latitude,elevation'.

Examples

data(simpleworld)
g <- globe(silicate::PATH(as(simpleworld, "SpatialLinesDataFrame")))
if (interactive()) {
 plot(g, lwd = 3)
 plot3d(g)
 rgl::open3d()
 wire3d(globe(DEL0(gebco*50)), col = "grey");
 rgl::spheres3d(0, 0, 0, rad = 6000000,col = "aliceblue")
}