Plot objects in OpenGL

Plot using the rgl-package.

# S3 method for trimesh
plot(x, ..., add = FALSE, add_normals = FALSE)

# S3 method for linemesh
plot(x, ..., add = FALSE)

# S3 method for pointmesh
plot(x, ..., add = FALSE)

Arguments

x
object from anglr

...
args for underlying plotting

add
add to existing plot if exists

add_normals
if TRUE use `rgl::addNormals` to smooth the model before plotting

Value

the rgl mesh3d object, invisibly

Details

The data structures from anglr are converted to their analogous forms used by the rgl package and plotted. These plot methods return the rgl form invisibly.