3D object plot

For SC edges are matched to their object/s. One object's properties is applied as colour. If color_ column is present on the data object table it is used.

# S3 method for SC
plot3d(x, ..., add = FALSE)

# S3 method for QUAD
plot3d(x, ..., add = FALSE)

# S3 method for PATH
plot3d(x, ..., add = FALSE)

# S3 method for sf
plot3d(x, ..., add = FALSE)

# S3 method for sfc
plot3d(x, ..., add = FALSE)

# S3 method for Spatial
plot3d(x, ..., add = FALSE)

# S3 method for trip
plot3d(x, ..., add = FALSE)

# S3 method for ARC
plot3d(x, ..., add = FALSE)

# S3 method for TRI
plot3d(x, ..., add = FALSE)

Arguments

x
silicate model, SC, TRI, ARC, or PATH

...

add

Value

vertices and segment indices, invisibly

Examples

library(silicate)
x <- SC(sf::read_sf(system.file("shape", "nc.shp", package = "sf")) %>%
   mutate(color_ = rainbow(100)))
plot3d(x); rglwidget()
worldz <- QUAD(gebco1)
## an easy way to exaggerate z is to reduce the radius of the globe
plot3d(globe(worldz, gproj = "+proj=geocent +a=10000"))

#> Error in array(x, c(length(x), 1L), if (!is.null(names(x))) list(names(x),     NULL) else NULL): 'data' must be of a vector type, was 'NULL'