Generate primitive-based spatial structures

Create primitive-based "mesh" structures from various inputs.

The 'rangl' package helps transcend the following problems in R and GIS generally:

rangl(x, ...)

# S3 method for SpatialLines
rangl(x, ...)

# S3 method for SpatialMultiPoints
rangl(x, ...)

# S3 method for SpatialPoints
rangl(x, ...)

# S3 method for SpatialPolygons
rangl(x, max_area = NULL, ...)

# S3 method for trip
rangl(x, ...)

# S3 method for mesh3d
rangl(x, ...)

Arguments

x

input data

...

arguments passed to methods

max_area

maximum area in coordinate system of x, passed to triangulate 'a' argument

Value

a list of tibble data frames, using the gris-map_table model

Details

#' Methods exist for SpatialPolygons, SpatialLines, rgl mesh3d(triangle) ...

coordinates beyond X and Y, or longitude and latitude

attributes stored where appropriate, on vertices, primitives, branches, or objects

topology and geometry are not conflated

spatial data can be properly represented as a graph of spatial primitives

polygons as true surfaces, not just glorified lines with a path-filling rule

TBD higher dimensional primitives are possible

TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction

I. Creation

rangl

create a rangl table set from various input types

II. Plotting

globe

convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form

plot.trimesh

plot 2D topology in 3D geometry space

plot.linemesh

plot 1D topology in 3D geometry space

plot.pointmesh

plot 0D topology in 3D geometry space

Examples

## -----------------------------------------------
## POLYGONS
library(maptools)
data(wrld_simpl)
b <- rangl(wrld_simpl)
plot(b)

Joining, by = "object_"

Joining, by = "triangle_"

if (require(rworldxtra)) {

data(countriesHigh)
sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
 "Indonesia", "Malaysia", "Fiji", "Australia")
a <- subset(countriesHigh, SOVEREIGNT %in% sv)
b7 <- rangl(a, max_area = 0.5)
plot(globe(b7))
}

Loading required package: rworldxtra

Joining, by = "object_"

Joining, by = "triangle_"

## -----------------------------------------------
## LINES
l1 <- rangl(as(a, "SpatialLinesDataFrame") )
plot(l1)

Joining, by = "object_"

Joining, by = "segment_"

plot(globe(l1))

Joining, by = "object_"

Joining, by = "segment_"

## -----------------------------------------------
## RGL TRIANGLES
library(rgl)

dod <- rangl(dodecahedron3d(col = "cyan"))
octo <- rangl(translate3d(octahedron3d(col = "blue"), 6, 0, 0))
plot(dod, col = viridis::viridis(5)[1], alpha = 0.3)

Joining, by = "object_"

Joining, by = "triangle_"

plot(octo, col = viridis::viridis(5)[5], alpha = 0.3)

Joining, by = "object_"

Joining, by = "triangle_"

bg3d("grey")