Draw a mesh as line segments in 3D

Draw line segments with rgl from any shape3d classed object. Produces a 3D scatterplot like that produced by rgl::plot3d(), but from a mesh-alike object.

# S3 method for sc
wire3d(x, ...)

# S3 method for TRI
wire3d(x, ...)

# S3 method for TRI0
wire3d(x, ...)

# S3 method for DEL
wire3d(x, ...)

# S3 method for DEL0
wire3d(x, ...)

# S3 method for QUAD
wire3d(x, ...)

# S3 method for matrix
wire3d(x, ...)

# S3 method for sf
wire3d(x, ...)

# S3 method for Spatial
wire3d(x, ...)

# S3 method for triangulation
wire3d(x, ...)

# S3 method for trip
wire3d(x, ...)

# S3 method for BasicRaster
wire3d(x, ...)

Arguments

x

sp, sf, raster, trip, or any other model understood by anglr/silicate

...

pass material3d properties to rgl

Details

Objects that are not explicitly surfaces will be triangulated in order to produce the mesh. Whether this is a good idea or not is an open question.

It is not currently technically defined or clear how colour properties are mapped to line segments by default ... there is a problem of what property to use from features that share the same vertex or edge, and we have put that aside and erred on the side of inaccuracy in favour of getting a pretty plot (hopefully). (Properties that come later - lower rows - win, I think.

See also