Convert object to a constrained-Delaunay triangulation

This relational-form Delaunay-based triangulation model is analogous to the 'TRI' model in the silicate package and formally extends the class of that model. A primitives-based shape-constrained triangulation. The Delaunay model is the mostly Delaunay scheme used by the provable-quality meshers.

DEL(x, ..., max_area = NULL)

# S3 method for default
DEL(x, ..., max_area = NULL)

# S3 method for PATH0
DEL(x, ..., max_area = NULL)

# S3 method for TRI
DEL(x, ..., max_area = NULL)

# S3 method for TRI0
DEL(x, ..., max_area = NULL)

# S3 method for SC
DEL(x, ..., max_area = NULL)

# S3 method for SC0
DEL(x, ..., max_area = NULL)

# S3 method for PATH
DEL(x, ..., max_area = NULL)

Arguments

x
input model

...
passed to the underlying Triangle library

max_area
the maximum area of a triangle

Value

DEL model

Details

Compare DEL() to its structural-form counterpart DEL0(). DEL() records a visible property on the triangle table, and this is queried by other functions for the status of triangles that belonged to a hole within a surface. These triangle are obviously useful, so they are kept but default to visible = FALSE and so are not plotted.

The Delaunay model is a constrained triangulation with a variety of constraint and qualification types. The Delaunay model has the odd but defining characteristic of not being always consistent with the Delaunay criterion. Edge inclusion is non-negotiable, but other constraints include (limit, or avoid) Steiner vertex insertion, a limit on the maximum area of a triangle, minimum triangle angle and strict adherence to the Delaunay criterion.

The Delaunay criterion forms the basis of this model, and is its defining characteristic without being strictly adhered to. This is awkward to describe but is the key property. From Wikipedia: The "Delaunay triangulation (also known as a Delone triangulation) for a given set P of discrete points in a plane is a triangulation DT(P) such that no point in P is inside the circumcircle of any triangle in DT(P). Delaunay triangulations maximize the minimum angle of all the angles of the triangles in the triangulation; they tend to avoid sliver triangles. . . . The Delaunay triangulation corresponds to the dual graph of the Voronoi diagram of P" https://en.wikipedia.org/wiki/Delaunay_triangulation.

This strict criterion is relaxed in small measure, to ensure that all edge-inputs are preserved and to allow further constraints such as triangle size and internal angle to be specified.

See also

Examples

plot3d(DEL(simpleworld))