Generate primitive-based spatial structures

Create primitive-based "mesh" structures from various inputs.

Colours not supported, this just gives the viridis palette sequentially.

anglr(x, ...)

# S3 method for SpatialMultiPoints
anglr(x, ...)

# S3 method for SpatialPoints
anglr(x, ...)

# S3 method for RasterLayer
anglr(x, z = x, na.rm = FALSE, ...)

# S3 method for trip
anglr(x, ...)

# S3 method for SpatialLines
anglr(x, ...)

# S3 method for SpatialPolygons
anglr(x, max_area = NULL, ...)

# S3 method for mesh3d
anglr(x, ...)

Arguments

x
input data

...
arguments passed to methods

z
raster, by default x is used

na.rm
remove missing values

max_area
maximum area in coordinate system of x, passed to triangulate 'a' argument

Value

a list of tibble data frames, using the gris-map_table model

quad_mesh

Details

#' Methods exist for SpatialPolygons, SpatialLines, rgl mesh3d(triangle) ...

Warning

anglr only checks for presence of triangle centres within known holes, so this doesn't pick up examples of overlapping areas e.g. https://github.com/hypertidy/anglr/issues/39

Examples

## -----------------------------------------------
## POLYGONS
library(maptools)
data(wrld_simpl)
b <- anglr(wrld_simpl)
plot(b)

#> Joining, by = "object_"

#> Joining, by = "triangle_"

#if (require(rworldxtra)) {

#data(countriesHigh)
#sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
#  "Indonesia", "Malaysia", "Fiji", "Australia")
#a <- subset(countriesHigh, SOVEREIGNT %in% sv)
#b7 <- anglr(a, max_area = 0.5)
#plot(globe(b7))
#}
## -----------------------------------------------
## LINES
#l1 <- anglr(as(a, "SpatialLinesDataFrame") )
#plot(l1)
#plot(globe(l1))

#data("flight_tracks", package = "silicate")
#r <- anglr(flight_tracks)
#plot(r)
#rgl::aspect3d(1, 1, 0.001)
#rgl::rglwidget()
library(raster)
w <- raster(volcano)
plot(anglr(w/300))

## -----------------------------------------------
## RGL TRIANGLES
library(rgl)

dod <- anglr(dodecahedron3d(col = "cyan"))
octo <- anglr(translate3d(octahedron3d(col = "blue"), 6, 0, 0))
plot(dod, col = viridis::viridis(5)[1], alpha = 0.3)

#> Joining, by = "object_"

#> Joining, by = "triangle_"

plot(octo, col = viridis::viridis(5)[5], alpha = 0.3)

#> Joining, by = "object_"

#> Joining, by = "triangle_"

bg3d("grey")