Tidy tables for topological spatial data structures.

The 'anglr' package show-cases getting past limitations in spatial data by extending models in the `silicate`` package:

coordinates beyond X and Y, or longitude and latitude

storing attributes on vertices, primitives, branches (parts), or objects

topology and geometry are properly separated

spatial data can be properly represented as a graph of spatial primitives

polygons as true surfaces, not just glorified lines with a path-filling rule

TBD higher dimensional primitives are possible

TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction

I. Creation

SC, PATH, TRI, and ARC silicate models are all supported, including the structural forms SC0, TRI0, PATH0	Spatial

most spatial types can be used directly	DEL

create a mostly-Delaunay shape-preserving constrained triangulation

II. Plotting

globe	convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form

plot3d.SC	plot 1D topology in 3D geometry space

plot3d.TRI	plot 2D topology in 3D geometry space (DEL or TRI)

III. Options and technicalities

There is an option set for the maximum number of triangles that can be generated by the DEL() or DEL0() models when using the max_area argument. Inspect the limit with getOption("anglr.max.triangles") or set a new limit with options(anglr.max.triangles = <new limit>).

In terms of the RTriangle::triangulate() function the max_area argument controls and masks the a argument for RTriangle. It's possible to pass in values for q, Y, j, D, S, V, and Q - but we don't recommend experimenting with these unless you know what they are for. There are no checks for the limit on the number of triangles that can be created and several of these do affect that.