Tidy tables for topological spatial data structures.

The 'anglr' package helps transcend the following general limitations:

coordinates beyond X and Y, or longitude and latitude

storing attributes on vertices, primitives, branches (parts), or objects

topology and geometry are properly separated

spatial data can be properly represented as a graph of spatial primitives

polygons as true surfaces, not just glorified lines with a path-filling rule

TBD higher dimensional primitives are possible

TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction

I. Creation

anglr 	 create a anglr table set from various input types

II. Plotting

globe 	 convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form

plot.trimesh 	 plot 2D topology in 3D geometry space

plot.linemesh 	 plot 1D topology in 3D geometry space

plot.pointmesh 	 plot 0D topology in 3D geometry space