Copy down values to vertices

Copy down provides ways to transfer object level data values to vertex level.

copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for mesh3d
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for SC
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for SC0
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for TRI
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for TRI0
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for PATH
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for ARC
copy_down(x, z = NULL, ..., .id = "z_")

# S3 method for QUAD
copy_down(x, z = NULL, ..., .id = "z_")

Arguments

x
a mesh3d or a silicate object

z
object specifying values to copy down, a vector of values, a column name, a raster (see details)

...
currently ignored

.id
character value, the name of the resulting column in the vertices, default is "z_"

Value

silicate model with vertex values copied to vertices

Details

Various methods are used depending on the second argument z.

If z is a raster (BasicRaster) a numeric value for each vertex is found by bilinear interpolation using raster::extract(raster, vertex, method = "bilinear"). Vertices are transformed into the space used by the raster if possible. (WIP ... Otherwise a warning is issued if there's not overlap ... WIP)

Beware, the CRS (coordinate reference system) space is fraught at the moment, proj4strings are out of favour, '+init=epsg:4326' doesn't work but 'EPSG:4326' does and sf will only give '$proj4string' dynamically, but it might be copied in the '$input' element, so the crs input lookup tries all these things (One day soon we'll just a crs and let PROJ sort it out, but for now we have to send all these messages around, it'll calm down in a few years (2024-ish)). If z is a character value, that column is taken from the object table.

The .id argument must be character and exist as a column name in the object table.

If z is a vector it's simply copied down.

No checking is done on the type of the result, and so there's nothing to stop the use of the recyling rule to expand out values, and nothing to stop the use of non numeric values being copied down. It's your model.

Examples

library(raster)

#> Warning: package 'raster' was built under R version 3.6.3

#> Loading required package: sp

#> Warning: package 'sp' was built under R version 3.6.3

r <- raster(volcano)
cl <- silicate::SC(rasterToContour(r))
plot3d(copy_down(cl, r))
## looks funny?
auto_3d(z = 15)

#> original axis lengths x,y,z:  0.9836066, 0.9885057,90.0985107

#> applying 'aspect3d(1.000, 1.000, 0.164)'