Generate primitive-based spatial structures

Create primitive-based "mesh" structures from various inputs.

Colours not supported, this just gives the viridis palette sequentially.

rangl(x, ...)

# S3 method for SpatialLines
rangl(x, ...)

# S3 method for SpatialMultiPoints
rangl(x, ...)

# S3 method for SpatialPoints
rangl(x, ...)

# S3 method for SpatialPolygons
rangl(x, max_area = NULL, ...)

# S3 method for RasterLayer
rangl(x, z = x, na.rm = FALSE, ...)

# S3 method for trip
rangl(x, ...)

# S3 method for mesh3d
rangl(x, ...)

Arguments

x

input data

...

arguments passed to methods

max_area

maximum area in coordinate system of x, passed to triangulate 'a' argument

z

raster, by default x is used

na.rm

remove missing values

Value

a list of tibble data frames, using the gris-map_table model

quad_mesh

Details

#' Methods exist for SpatialPolygons, SpatialLines, rgl mesh3d(triangle) ...

Examples

## -----------------------------------------------
## POLYGONS
library(maptools)
data(wrld_simpl)
b <- rangl(wrld_simpl)
plot(b)

Joining, by = "object_"

Joining, by = "triangle_"

#if (require(rworldxtra)) {

#data(countriesHigh)
#sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
#  "Indonesia", "Malaysia", "Fiji", "Australia")
#a <- subset(countriesHigh, SOVEREIGNT %in% sv)
#b7 <- rangl(a, max_area = 0.5)
#plot(globe(b7))
#}
## -----------------------------------------------
## LINES
#l1 <- rangl(as(a, "SpatialLinesDataFrame") )
#plot(l1)
#plot(globe(l1))
library(raster)
w <- raster(volcano)
plot(rangl(w/300))

## -----------------------------------------------
## RGL TRIANGLES
library(rgl)

dod <- rangl(dodecahedron3d(col = "cyan"))
octo <- rangl(translate3d(octahedron3d(col = "blue"), 6, 0, 0))
plot(dod, col = viridis::viridis(5)[1], alpha = 0.3)

Joining, by = "object_"

Joining, by = "triangle_"

plot(octo, col = viridis::viridis(5)[5], alpha = 0.3)

Joining, by = "object_"

Joining, by = "triangle_"

bg3d("grey")